we need more details and better title for this document
It's a valuable document and very useful, we need more details and better title for this document that covers what customer will get out of it in simpler terms. the main reason is customer and support couldn't locate it is the title "portal vanity domain" we can use something more straightforward like "How to customize Exchange Portal DNS record" and then add little details to the top of the document.  
Related support ticket:
https://na29.salesforce.com/50034000018IaPV